refactor: move main logic into MainController function
- Application.kt: removed logic, now only calls MainController
- MainController.kt: moved main logic into this function
