+++ src/main/kotlin/racingcar/Application.kt
@@ -1,19 +1,8 @@
 package racingcar
 
-import racingcar.view.InputView
-import racingcar.view.OutputView
-import racingcar.model.findTheWinnersIndex
+import racingcar.controller.mainController
 
-// TO-DO: Refactor / move to mainController
 fun main() {
-    val splitCarNames = InputView().inputCarNames()
-    val theNumberOfRounds = InputView().inputTheNumberOfRounds()
-    OutputView().printCarRacingStartMessage()
-
-    val finalResultPregressList = OutputView().printCarNamesAndProcess(splitCarNames, theNumberOfRounds)
-    val maxIntIndex = findTheWinnersIndex(finalResultPregressList)
-    OutputView().printWinners(splitCarNames, maxIntIndex)
+    mainController()
 }
-
-
